Sort SignedAttrs in AddSignerInfo().
X690 Section 11.6 specifies a required order for SET OFs in DER.
Go 1.15 updated encoding/asn1 to perform this sorting, which caused
a discrepancy between the order of the attributes in SignerInfo
and the order in the marshaled form generated by MarshaledForSigning().

The attributes are now sorted before calling MarshaledforSigning()
to ensure consistency. This fixes interoperability with other systems
(e.g. OpenSSL CMS) for signatures generated by ietf-cms when compiled
with Go >= 1.15.